fix(structure): finalized file refactoring.
