+++ src/app/components/work-item-comment/work-item-comment.component.spec.ts
@@ -50,7 +50,7 @@ import { ModalModule } from 'ngx-modal';
 import { AreaModel } from '../../models/area.model';
 import { AreaService } from '../../services/area.service';
 import { CollaboratorService } from '../../services/collaborator.service';
-import { DynamicFieldComponent } from '../dynamic-form/dynamic-field.component';
+import { DynamicFieldComponent } from '../dynamic-field/dynamic-field.component';
 import { TypeaheadDropdown } from '../typeahead-dropdown/typeahead-dropdown.component';
 import { IterationModel } from '../../models/iteration.model';
 import { IterationService } from '../../../services/iteration.service';
