+++ src/app/components/work-item-create/work-item-create-selector/work-item-create-selector.component.scss
