+++ src/app/components/work-item-create/work-item-create.component.html
@@ -1,0 +1,1 @@
+<create-selector-widget #detailAddTypeSelector [workItemTypes]="workItemTypes" (onSelect)="onChangeType($event)"></create-selector-widget>
