--- src/app/components/work-item-detail-add-type-selector/work-item-detail-add-type-selector.component.html
@@ -1,1 +1,0 @@
-<detail-add-type-selector-widget #detailAddTypeSelector [workItemTypes]="workItemTypes" (onSelect)="onChangeType($event)"></detail-add-type-selector-widget>
