+++ src/app/components/work-item-iteration-modal/work-item-iteration-modal.module.ts
@@ -12,7 +12,7 @@ import {
 import { ModalModule } from 'ngx-modal';
 import {
   FabPlannerAssociateIterationModalComponent
-} from './work-item-iteration-association-modal.component';
+} from './work-item-iteration-modal.component';
 
 @NgModule({
   imports: [
